Move pgUnionAll pagination to runtime
